нормализация

+++ src/main/java/ru/marilka/swotbackend/model/request/ParticipantDto.java
@@ -2,5 +2,5 @@ package ru.marilka.swotbackend.model.request;
 
 public record ParticipantDto(
         Long userId,
-        double coefficient
+        int coefficient
 ) {}
